+++ src/main/java/com/better/CommuteMate/category/application/dto/request/PutManagerCategoryRequest.java
@@ -7,7 +7,7 @@ import java.util.List;
 public record PutManagerCategoryRequest(
 
         @Schema(description = "매니저 ID", example = "5")
-        Integer managerId,
+        Long managerId,
 
         @Schema(description = "새로 매핑할 category 이름 리스트", example = "[\"도서관시스템\", \"학사정보시스템\"]")
         List<String> categoryNames
